requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq===0.1.9
unstructured==0.14.6
selenium>=4.9.0
chromadb==0.5.0
streamlit>=1.38.0
pandas>=2.0.0
python-dotenv==1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-jobspy
sentence-transformers
langchain-huggingface 
langchain-chroma